requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
exceptiongroup==1.3.0
fastapi==0.121.3
idna==3.11
mysql-connector-python==9.5.0
pydantic==2.12.4
pydantic_core==2.41.5
sniffio==1.3.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
